Move unit tests to its own directory
