+++ packages/protocol/test/uint/libs/LibBlockHeader.test.ts
@@ -1,5 +1,5 @@
 import { expect } from "chai"
-import * as log from "../../tasks/log"
+import * as log from "../../../tasks/log"
 const hre = require("hardhat")
 const ethers = hre.ethers
 const EBN = ethers.BigNumber
